->98std

+++ src/config/Config.cpp
@@ -18,7 +18,7 @@ Config::~Config()
 
 void Config::fillLocations()
 {
-    for (int i = 0; i < _locations.size(); i++)
+    for (unsigned long i = 0; i < _locations.size(); i++)
     {
         if (_locations[i].second->_autoindex.empty())
             _locations[i].second->_autoindex = this->_autoindex;
@@ -83,26 +83,26 @@ void Config::printConfig() const {
     std::cout<< "_host " << _host << std::endl;
     std::cout<< "_port " << _port << std::endl;
     std::cout << "_server_name: ";
-    for (int i = 0; i < this->_server_name.size(); i++)
+    for (unsigned long i = 0; i < this->_server_name.size(); i++)
         std::cout << this->_server_name[i] << " ";
     std::cout << "\n_autoIndex: " << this->_autoindex << std::endl;
     std::cout << "_index: ";
-    for (int j = 0; j < this->_index.size(); j++) {
+    for (unsigned long j = 0; j < this->_index.size(); j++) {
             std::cout << this->_index[j] << " "; 
     }
     std::cout << "\n_root: " << this->_root << std::endl;
     std::cout << "_upload_path: " << this->_upload_path << std::endl;
     std::cout << "_client_max_body_size: " << this->_client_max_body_size << std::endl;
-    for (int j = 0; j < this->_error_page.size(); j++)
+    for (unsigned long j = 0; j < this->_error_page.size(); j++)
     {
         std::cout << "_error_page: (" << this->_error_page[j].first << ", " << this->_error_page[j].second << ")" << std::endl;
     }
     std::cout << "_methods: ";
-    for (int j = 0; j < this->_methods.size(); j++)
+    for (unsigned long j = 0; j < this->_methods.size(); j++)
         std::cout << this->_methods[j] << " ";
 
     std::cout << std::endl;
-    for (int j = 0; j < this->_cgi.size(); j++)
+    for (unsigned long j = 0; j < this->_cgi.size(); j++)
     {
         std::cout << "_cgi: (" << this->_cgi[j].first << ", " << this->_cgi[j].second << ")" << std::endl;
     }
